Module aws_sdk_route53recoverycontrolconfig::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

403 response - AccessDeniedException. You do not have sufficient access to perform this action.

ConflictException

409 response - ConflictException.

CreateClusterError

CreateControlPanelError

CreateRoutingControlError

CreateSafetyRuleError

DeleteClusterError

DeleteControlPanelError

DeleteRoutingControlError

DeleteSafetyRuleError

DescribeClusterError

DescribeControlPanelError

DescribeRoutingControlError

DescribeSafetyRuleError

InternalServerException

500 response - InternalServiceError. Temporary service error. Retry the request.

ListAssociatedRoute53HealthChecksError

ListClustersError

ListControlPanelsError

ListRoutingControlsError

ListSafetyRulesError

ResourceNotFoundException

404 response - The query string contains a syntax error or resource not found.

ServiceQuotaExceededException

402 response

ThrottlingException

429 response - ThrottlingException.

UpdateControlPanelError

UpdateRoutingControlError

UpdateSafetyRuleError

ValidationException

400 response - Multiple causes. For example, you might have a malformed query string and input parameter might be out of range, or you used parameters together incorrectly.

Enums

CreateClusterErrorKind

CreateControlPanelErrorKind

CreateRoutingControlErrorKind

CreateSafetyRuleErrorKind

DeleteClusterErrorKind

DeleteControlPanelErrorKind

DeleteRoutingControlErrorKind

DeleteSafetyRuleErrorKind

DescribeClusterErrorKind

DescribeControlPanelErrorKind

DescribeRoutingControlErrorKind

DescribeSafetyRuleErrorKind

ListAssociatedRoute53HealthChecksErrorKind

ListClustersErrorKind

ListControlPanelsErrorKind

ListRoutingControlsErrorKind

ListSafetyRulesErrorKind

UpdateControlPanelErrorKind

UpdateRoutingControlErrorKind

UpdateSafetyRuleErrorKind