Module aws_sdk_route53recoverycontrolconfig::model

Modules

assertion_rule

See AssertionRule

assertion_rule_update

See AssertionRuleUpdate

cluster

See Cluster

cluster_endpoint

See ClusterEndpoint

control_panel

See ControlPanel

gating_rule

See GatingRule

gating_rule_update

See GatingRuleUpdate

new_assertion_rule

See NewAssertionRule

new_gating_rule

See NewGatingRule

routing_control

See RoutingControl

rule

See Rule

rule_config

See RuleConfig

Structs

AssertionRule

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

AssertionRuleUpdate

An update to an assertion rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

Cluster

A cluster is a set of five consensus-forming Regional endpoints that represent the infrastructure that hosts your routing controls. Typically, you host together on one cluster all of the routing controls for your applications.

ClusterEndpoint

A cluster endpoint. Specify an endpoint when you want to set or retrieve a routing control state in the cluster.

ControlPanel

A control panel represents a group of routing controls that can be changed together in a single transaction.

GatingRule

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

GatingRuleUpdate

Update to a gating rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

NewAssertionRule

A new assertion rule for a control panel.

NewGatingRule

A new gating rule for a control panel.

RoutingControl

A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control traffic routing.

Rule

A safety rule. A safety rule can be an assertion rule or a gating rule.

RuleConfig

The rule configuration for an assertion rule. That is, the criteria that you set for specific assertion controls (routing controls) that specify how many controls must be enabled after a transaction completes.

Enums

RuleType

An enumerated type that determines how the evaluated rules are processed. RuleType can be one of the following:

ATLEAST - At least N routing controls must be set. You specify N as the Threshold in the rule configuration.

AND - All routing controls must be set. This is a shortcut for "At least N," where N is the total number of controls in the rule.

OR - Any control must be set. This is a shortcut for "At least N," where N is 1.

Status

The deployment status of a resource. Status can be one of the following:

PENDING: Amazon Route 53 Application Recovery Controller is creating the resource.

DEPLOYED: The resource is deployed and ready to use.

PENDING_DELETION: Amazon Route 53 Application Recovery Controller is deleting the resource.